refactor(client): extract @kolu/solid-ui presentation atoms
Move the design-system atoms (Toggle, Kbd, SegmentedControl, Row, Section,
Surface, stackLayers, Tip) into a reusable workspace package. All eight are
fully self-contained (solid-js + @corvu/tooltip only). Subpath exports keep
each module's default/named contract, so the 36 import sites are a pure
path rewrite.

Behaviour-preserving. client/src: 24,942 -> 24,662 LOC (-280).
just check + just test-unit green (200 client tests).
