+++ packages/client/src/EmptyState.tsx
@@ -5,9 +5,9 @@ import { terminalKey } from "kolu-common/terminalKey";
 import { type Component, createMemo, createSignal, For, Show } from "solid-js";
 import { ACTIONS } from "./input/actions";
 import { formatKeybind } from "./input/keyboard";
-import Kbd from "./ui/Kbd";
-import { surface } from "./ui/Surface";
-import Toggle from "./ui/Toggle";
+import Kbd from "@kolu/solid-ui/Kbd";
+import { surface } from "@kolu/solid-ui/Surface";
+import Toggle from "@kolu/solid-ui/Toggle";
 
 const chrome = surface();
