+++ packages/client/src/MobileChromeSheet.tsx
@@ -23,7 +23,7 @@ import { useRightPanel } from "./right-panel/useRightPanel";
 import type { WsStatus } from "./rpc/rpc";
 import SettingsPopover from "./settings/SettingsPopover";
 import { InspectorToggleIcon, SettingsIcon } from "@kolu/solid-icons";
-import Kbd from "./ui/Kbd";
+import Kbd from "@kolu/solid-ui/Kbd";
 
 const statusStyles: Record<WsStatus, string> = {
   connecting: "bg-warning animate-pulse",
