+++ packages/client/src/recorder/RecordPopover.tsx
@@ -4,8 +4,8 @@
 
 import { type Component, createEffect, For, Show } from "solid-js";
 import { Portal } from "solid-js/web";
-import { surface } from "../ui/Surface";
-import Toggle from "../ui/Toggle";
+import { surface } from "@kolu/solid-ui/Surface";
+import Toggle from "@kolu/solid-ui/Toggle";
 import { useAnchoredPopover } from "../ui/useAnchoredPopover";
 import LevelMeter from "./LevelMeter";
 import { useRecorder } from "./useRecorder";
