+++ packages/client/src/right-panel/MetadataInspector.tsx
@@ -14,8 +14,8 @@ import {
   stateLabels,
 } from "../ui/agentDisplay";
 import { PrStateIcon, TerminalIcon, WorktreeIcon } from "@kolu/solid-icons";
-import Row from "../ui/Row";
-import Section from "../ui/Section";
+import Row from "@kolu/solid-ui/Row";
+import Section from "@kolu/solid-ui/Section";
 
 const MetadataInspector: Component<{
   meta: TerminalMetadata | null;
