+++ packages/client/src/right-panel/ModeChipPicker.tsx
@@ -15,7 +15,7 @@ import type { CodeTabView } from "kolu-common/surface";
 import { type Component, createMemo, createSignal, For, Show } from "solid-js";
 import { Dynamic, Portal } from "solid-js/web";
 import { ChevronDownIcon } from "@kolu/solid-icons";
-import { surface } from "../ui/Surface";
+import { surface } from "@kolu/solid-ui/Surface";
 import { useAnchoredPopover } from "../ui/useAnchoredPopover";
 
 export type ModeOption = {
