+++ packages/client/src/terminal/SearchBar.tsx
@@ -16,8 +16,8 @@ import {
   Show,
 } from "solid-js";
 import { ChevronDownIcon, ChevronUpIcon, CloseIcon } from "@kolu/solid-icons";
-import { surface } from "../ui/Surface";
-import Tip from "../ui/Tip";
+import { surface } from "@kolu/solid-ui/Surface";
+import Tip from "@kolu/solid-ui/Tip";
 
 const SEARCH_OPTIONS: ISearchOptions = {
   incremental: true,
