+++ packages/client/src/terminal/TerminalMeta.tsx
@@ -18,7 +18,7 @@ import { IntentMarkdownInline } from "../intent/IntentMarkdown";
 import { annotationLine } from "../intent/text";
 import { agentWorkflow } from "../ui/agentDisplay";
 import { PrStateIcon, WorktreeIcon } from "@kolu/solid-icons";
-import Tip from "../ui/Tip";
+import Tip from "@kolu/solid-ui/Tip";
 import ChecksIndicator from "./ChecksIndicator";
 import { PrUnavailableButton } from "./PrUnavailablePopover";
 import { prTooltip } from "./prTooltip";
