+++ packages/client/src/ui/CodeContextMenu.tsx
@@ -11,7 +11,7 @@ import { Dynamic, Portal } from "solid-js/web";
 import { toast } from "solid-sonner";
 import { match } from "ts-pattern";
 import { writeTextToClipboard } from "./clipboard";
-import { surface } from "./Surface";
+import { surface } from "@kolu/solid-ui/Surface";
 
 /** Two verbs over the same selection noun: copy a string to the clipboard,
  *  or invoke an action callback. The discriminator keeps the dispatch
