+++ packages/solid-ui/src/SegmentedControl.tsx
