+++ packages/solid-ui/src/Tip.tsx
